--- backend/src/main/java/vn/viettel/quanlycongno/service/SecurityService.java
@@ -1,37 +1,0 @@
-package vn.viettel.quanlycongno.service;
-
-import lombok.AllArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import vn.viettel.quanlycongno.entity.Staff;
-import vn.viettel.quanlycongno.repository.StaffRepository;
-
-@Service
-@AllArgsConstructor
-public class SecurityService {
-    private final StaffRepository staffRepository;
-
-    public String getCurrentUsername() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (authentication == null || !authentication.isAuthenticated()) {
-            throw new AccessDeniedException("User not authenticated");
-        }
-        return authentication.getName();
-    }
-
-    public Staff getCurrentStaff() {
-        String username = getCurrentUsername();
-        return staffRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + username));
-    }
-
-    public boolean isAdmin() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        return authentication.getAuthorities().stream()
-                .anyMatch(a -> a.getAuthority().equals("ROLE_ADMIN"));
-    }
-}
