+++ backend/src/main/java/vn/viettel/quanlycongno/service/StaffService.java
@@ -1,20 +1,13 @@
 package vn.viettel.quanlycongno.service;
 
-import lombok.AllArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import vn.viettel.quanlycongno.repository.StaffRepository;
 
 @Service
-@AllArgsConstructor
-public class StaffService implements UserDetailsService {
-    private final StaffRepository staffRepository;
+public interface StaffService extends UserDetailsService {
 
     @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return staffRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + username));
-    }
+    UserDetails loadUserByUsername(String username) throws UsernameNotFoundException;
 }
